feat: Reverted previous commit, asynchronously started connection to peers
