+++ crates/libtortillas/Cargo.toml
@@ -21,7 +21,6 @@ tracing = "0.1.41"
 async-trait = "0.1.88"
 librqbit-utp = "0.2.3"
 bitvec = "1.0.1"
-futures = "0.3.31"
 
 [dev-dependencies]
 tracing-test = "0.2.5"
